chore(textfield): move dependency components to components folder and make csf
