+++ src/components/Label/index.ts
