+++ src/index.ts
@@ -22,6 +22,7 @@ export { default as Dropdown } from './components/Dropdown';
 export { default as DropdownButton } from './components/DropdownButton';
 export { default as DropdownMenu } from './components/DropdownMenu';
 export { default as DropdownMenuItem } from './components/DropdownMenuItem';
+export { default as FieldNote } from './components/FieldNote';
 export { default as Fieldset } from './components/Fieldset';
 export { default as FieldsetItems } from './components/FieldsetItems';
 export { default as FieldsetLegend } from './components/FieldsetLegend';
@@ -33,6 +34,7 @@ export { default as HorizontalStepper } from './components/HorizontalStepper';
 export { default as Hr } from './components/Hr';
 export { default as Icon } from './components/Icon';
 export { default as InlineNotification } from './components/InlineNotification';
+export { default as Label } from './components/Label';
 export { default as Layout } from './components/Layout';
 export { default as LayoutContainer } from './components/LayoutContainer';
 export { default as LayoutLinelengthContainer } from './components/LayoutLinelengthContainer';
@@ -51,6 +53,7 @@ export { default as Tabs } from './components/Tabs';
 export { default as Tag } from './components/Tag';
 export { default as Text } from './components/Text';
 export { default as TextField } from './components/TextField';
+export { default as TextInput } from './components/TextInput';
 export { default as Toast } from './components/Toast';
 export { default as Toolbar } from './components/Toolbar';
 export { default as ToolbarItem } from './components/ToolbarItem';
