--- src/upcoming-components/Label/Label.stories.tsx
@@ -1,30 +1,0 @@
-import { BADGE } from '@geometricpanda/storybook-addon-badges';
-import { Story, Meta } from '@storybook/react';
-import React from 'react';
-
-import { Label, Props } from './Label';
-
-export default {
-  title: 'Atoms/Forms/Label',
-  component: Label,
-  parameters: {
-    badges: [BADGE.BETA],
-  },
-} as Meta;
-
-const Template: Story<Props> = (args) => <Label {...args} />;
-
-const InvertedTemplate: Story<Props> = (args) => (
-  <div style={{ padding: '1rem', background: '#000' }}>
-    <Label {...args} />
-  </div>
-);
-
-export const Default = Template.bind({});
-Default.args = { text: 'Label', required: true };
-
-export const Optional = Template.bind({});
-Optional.args = { text: 'Label', required: false };
-
-export const Inverted = InvertedTemplate.bind({});
-Inverted.args = { inverted: true, text: 'Label', required: true };
